Project refactor based onion dependencies
