+++ ApiApplication/ApiApplication.csproj
@@ -15,6 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Lodgify.Cinema.DomainService\Lodgify.Cinema.DomainService.csproj" />
     <ProjectReference Include="..\Lodgify.Cinema.Infrastructure.Ioc\Lodgify.Cinema.Infrastructure.Ioc.csproj" />
   </ItemGroup>
