+++ ApiApplication/ApiApplication.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ApiApplication", "ApiApplication.csproj", "{979FF842-D59A-44BD-A726-4850A4B9647A}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A7B7247A-2A34-442B-ABD6-F2053B6B3C24}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "0 - Solution Items", "0 - Solution Items", "{A7B7247A-2A34-442B-ABD6-F2053B6B3C24}"
 	ProjectSection(SolutionItems) = preProject
 		..\cUrls.txt = ..\cUrls.txt
 		..\Feedback.md = ..\Feedback.md
@@ -29,6 +29,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lodgify.Cinema.Infrastructure.Data", "..\Lodgify.Cinema.Data\Lodgify.Cinema.Infrastructure.Data.csproj", "{775C9774-2F4E-4FDF-998B-6F3ACD28556C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lodgify.Cinema.Infrastructure.Ioc", "..\Lodgify.Cinema.Infrastructure.Ioc\Lodgify.Cinema.Infrastructure.Ioc.csproj", "{19F0A16B-3E45-4D7A-AB76-51852A77F020}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lodgify.Cinema.DomainService", "..\Lodgify.Cinema.DomainService\Lodgify.Cinema.DomainService.csproj", "{8FE9DCAD-5614-4B56-820E-4F7FFF136B26}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -56,6 +58,10 @@ Global
 		{19F0A16B-3E45-4D7A-AB76-51852A77F020}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{19F0A16B-3E45-4D7A-AB76-51852A77F020}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{19F0A16B-3E45-4D7A-AB76-51852A77F020}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8FE9DCAD-5614-4B56-820E-4F7FFF136B26}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8FE9DCAD-5614-4B56-820E-4F7FFF136B26}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8FE9DCAD-5614-4B56-820E-4F7FFF136B26}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8FE9DCAD-5614-4B56-820E-4F7FFF136B26}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -66,6 +72,7 @@ Global
 		{1D95BF94-77FB-431A-87DF-A0146DE349D4} = {F61633AC-0A5B-4BCE-B970-E99CD2A4BF4B}
 		{775C9774-2F4E-4FDF-998B-6F3ACD28556C} = {72390D34-8A83-4F31-9217-6CB82E4E6FC3}
 		{19F0A16B-3E45-4D7A-AB76-51852A77F020} = {72390D34-8A83-4F31-9217-6CB82E4E6FC3}
+		{8FE9DCAD-5614-4B56-820E-4F7FFF136B26} = {75667C0F-5DD2-4DD8-8B4D-460ED799DC6D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C52A7E7E-4F02-4B44-B7D4-995DE72CF0B3}
