+++ ApiApplication/Startup.LogConfiguration.cs
@@ -1,5 +1,4 @@
-﻿using Lodgify.Cinema.Domain.Notification;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
 namespace ApiApplication
@@ -7,8 +6,6 @@ namespace ApiApplication
     public static class LogConfiguration
     {
         public static IServiceCollection ConfigureNotificationAndLog(this IServiceCollection services) =>
-             services.AddScoped<ILogger, Logger<string>>()
-                     .AddScoped<IDomainNotification, DomainNotification>();
-
+             services.AddScoped<ILogger, Logger<string>>();
     }
 }
