--- Lodgify.Cinema.Domain/Class1.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Lodgify.Cinema.Domain
-{
-    public class Class1
-    {
-
-    }
-}
