+++ Lodgify.Cinema.Domain/Contract/Notification/IDomainNotification.cs
