+++ Lodgify.Cinema.Domain/Contract/Pagination/IPaginatedRequest.cs
