+++ Lodgify.Cinema.Infrastructure.Ioc/IocConfiguration.cs
@@ -1,10 +1,7 @@
 ﻿using Lodgify.Cinema.Domain.Contract.Repositorie;
-using Lodgify.Cinema.Infrastructure.Data.Context;
-using Lodgify.Cinema.Infrastructure.Data.Database;
+using Lodgify.Cinema.Domain.Notification;
+using Lodgify.Cinema.DomainService.Notification;
 using Lodgify.Cinema.Infrastructure.Data.Repositorie;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Lodgify.Cinema.Infrastructure.Ioc
@@ -13,28 +10,9 @@ namespace Lodgify.Cinema.Infrastructure.Ioc
     {
         public static IServiceCollection ConfigureIocBusinessDependencies(this IServiceCollection services)
         {
-            services.AddTransient<IShowtimesRepository, ShowtimesRepository>();
+            services.AddTransient<IShowtimesRepository, ShowtimesRepository>()
+                    .AddScoped<IDomainNotification, DomainNotification>();
             return services;
-        }
-
-
-        public static IServiceCollection ConfigureIocDbDependencies(this IServiceCollection services)
-        {
-            services.AddDbContext<CinemaContext>(options =>
-            {
-                options.UseInMemoryDatabase("CinemaDb")
-                    .EnableSensitiveDataLogging()
-                    .ConfigureWarnings(b => b.Ignore(InMemoryEventId.TransactionIgnoredWarning));
-            });
-
-            return services;
-        }
-
-        public static void Initialize(this IApplicationBuilder app)
-        {
-            using var serviceScope = app.ApplicationServices.GetRequiredService<IServiceScopeFactory>().CreateScope();
-            var context = serviceScope.ServiceProvider.GetService<CinemaContext>();
-            SampleData.Initialize(context);
         }
     }
 }
